[LIVY-572] Avoid usage of spark classes in ColumnBuffer
